src/main/java/com/upendiproject/localanimalapi/controller/BaseController.java
package com.upendiproject.localanimalapi.controller;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.servlet.NoHandlerFoundException;

import javax.servlet.http.HttpServletRequest;

@Controller
public class BaseController {

    private static final Logger logger = LoggerFactory.getLogger(BaseController.class);

    @ExceptionHandler(Exception.class)
    @ResponseStatus(value= HttpStatus.INTERNAL_SERVER_ERROR)
    public ResponseEntity handleException(HttpServletRequest req, Exception ex) {
        logger.error("Method: handleException - An exception has occurred...", ex);
        return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
    }

    @ExceptionHandler(RuntimeException.class)
    @ResponseStatus(value= HttpStatus.INTERNAL_SERVER_ERROR)
    public ResponseEntity handleRuntimeException(HttpServletRequest req, Exception ex) {
        logger.error("Method: handleException - An exception has occurred...", ex);
        return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
    }

    @ExceptionHandler(NoHandlerFoundException.class)
    @ResponseStatus(value= HttpStatus.NOT_FOUND)
    public ResponseEntity handle404Exception(HttpServletRequest req, Exception ex) {
        logger.error("Method: handleException - An exception has occurred...", ex);
        return new ResponseEntity(HttpStatus.NOT_FOUND);
    }

}
